// front-end/src/components/atoms/Typography/interface.ts
export type IHTMLTag = 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'p' | 'span' | 'div';
export type ITextSize = 'xs' | 'sm' | 'base' | 'lg' | 'xl' | '2xl' | '3xl' | '4xl' | '5xl' | '6xl' | '7xl' | '8xl' | '9xl';
export type ITextWeight = 'thin' | 'extralight' | 'light' | 'normal' | 'medium' | 'semibold' | 'bold' | 'extrabold' | 'black';
export type ITextAlign = 'left' | 'center' | 'right';
export type IColorShade = '50' | '100' | '200' | '300' | '400' | '500' | '600' | '700' | '800' | '900';
export type IColor =
  | 'slate'
  | 'gray'
  | 'sky'
  | 'zinc'
  | 'neutral'
  | 'stone'
  | 'red'
  | 'orange'
  | 'amber'
  | 'yellow'
  | 'lime'
  | 'green'
  | 'emerald'
  | 'teal'
  | 'cyan'
  | 'blue'
  | 'indigo'
  | 'violet'
  | 'purple'
  | 'fuchsia'
  | 'pink'
  | 'rose';
export type IBreakWords = 'normal' | 'words' | 'all';
export type IStyleType = 'headline' | 'subHeadline' | 'cardTitle';
export type HTMLTag = 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'p' | 'span' | 'div';

export type TextAlign = 'left' | 'center' | 'right' | 'justify';

export type BreakWords = 'normal' | 'words' | 'all';

export type VariantType = 'h1' | 'h2' | 'h3' | 'h4' |"h5" | "h6" |'cardTitle' | 'paragraph' | 'heroTitle' | "smallParagraph" |'small';

export type TextTransform = 'capitalize' | 'lowercase' | 'uppercase';

export type FontWeight = 'regular' | 'medium' | 'semibold' | 'bold' | 'extrabold';

export type FontStyle = 'italic' | 'oblique';

export type TextColor = 'primary' | 'secondary' | 'gray' | 'green' | 'orange' | 'white' | 'red' | 'slate';